+++ src/workspace.ts
@@ -1,6 +1,6 @@
 import '../styles/main.scss';
 
-export { AbortScope, mapAbortedToNull, delay } from './coreUtils/async';
+export { mapAbortedToNull, delay } from './coreUtils/async';
 export { moveComparator, shallowArrayEqual } from './coreUtils/collections';
 export { useColorScheme } from './coreUtils/colorScheme';
 export {